requirements.txt
absl-py==0.13.0
altair==4.1.0
altgraph==0.17.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.1
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.2
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyler==0.2.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
streamlit==1.0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
toml==0.10.2
toolz==0.11.1
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.0
traitlets==5.1.0
typing-extensions==3.10.0.0
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2